Fix Safari bug with flickering elements on drag
We can not use "event.relatedTarget" to check if the native element still
contains the dragged element because of a Safari bug with "NULL" as
"event.relatedTarget" value on the dragleave event.

See: https://stackoverflow.com/questions/68745781/dragevent-relatedtarget-returns-null-in-safari

The solution is to use a counter instead of the related target that decrease
on drag leave and increase on drag enter.